# src/express/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: Memora API
  description: Memora openApi 3.0.1 specification
  version: 1.0.23
  contact:
    email: to@mail.io
servers:
  - description: Localhost
    url: http://localhost:3001
...
paths:
...
  /users/inbox-summary:
    get:
      tags:
        - User
      summary: Returns inbox summary for the current user
      operationId: getInboxSummaryCtr
      security:
        - userLoginBearerAuth: []
      responses:
        '200':
          description: Inbox summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 5
                  deskSub:
                    type: string
                    format: uuid
  /users/{nickname}/public:
    get:
      tags:
        - User
      summary: Returns a public user profile
      operationId: getPublicProfileCtr
      parameters:
        - name: nickname
          in: path
          required: true
          schema:
            type: string
            pattern: '^[a-z0-9_]{3,50}$'
      responses:
        '200':
          description: Public profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  nickname:
                    type: string
                  memberSince:
                    type: string
                    format: date-time
                  desks:
                    type: array
                    items:
                      type: object
                      properties:
                        sub:
                          type: string
                          format: uuid
                        title:
                          type: string
                        cardCount:
                          type: integer
                        totalSaves:
                          type: integer
        '404':
          description: User not found
        '422':
          description: Invalid nickname
...
  /games/add-to-inbox:
    post:
      tags:
        - Review
      summary: Saves a feed card to the user inbox desk
      operationId: addCardToInboxCtr
      security:
        - userLoginBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - cardSub
              properties:
                cardSub:
                  type: string
                  format: uuid
      responses:
        '204':
          description: Card saved to inbox
...
components:
...
  securitySchemes:
    userLoginBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT